Issue #377 - use the VMUSB drivers in place rather than as copies
Import the NADC2530 support directly from vmusb/daqconfig
